Add simple inter-process communication
This IPC sends binary data over pipes. It relies on both ends using the
same size and encoding for types. It will be used to communicate between
different processes of the same executable, so they will have the same
size and encoding, since it is the same executable running on the same
computer.