FEATURE: Track boot loader on install
On install, if LICK exists and the bootloader is installed, use the old
LICK to uninstall the bootloader since it is compatible; then use the
new LICK to install the bootloader again.

This has two benefits. The first is that the technique used to install
the boot loader can be updated without needing to support uninstalling
old methods. Second, the files used in the bootloader are updated.