FEATURE: GUI is not supported on 9X
Windows 95 is missing some symbols (related to using C++) to run FLTK.
Even if this was patched, Windows 98 doesn't show the window, so it
probably won't work in Windows 95 either.

I don't believe there is enough demand for a Windows 95/98 GUI to
warrant creating it, especially since the CLI isn't too hard to
use. If there is demand, I'll reconsider.

This commit stops lick-fltk from hanging on Windows 98, and simply
redirects the user to the CLI.