FEATURE: Support Windows ME
By default, Windows ME ignores config.sys, which LICK depends on. To fix
this, ask the user to install Me2Dos, a program which patches the system
to read config.sys.

I do not believe there is enough demand to warrant automating the
running of winmedos.com, especially since it doesn't provide a "nice"
command line interface. Additionally, this allows the user to pick a
different install method if winmedos fails to patch the system, or if
it was already patched by a different method.