FEATURE: Use Windows API for drag and drop
The way recent Windows versions work, drag and drop can only occur between
two processes that have the same privileges. This is the cause of a lot
of complexity in the GUI, especially around message passing between the
frontend and backend. Thus, working around this would simplify the GUI.

It is possible to unfilter the drag and drop messages, but this
workaround doesn't work with the way FLTK implements drag and drop. This
commit converts the drag and drop handler to a custom one, which can use
the workaround.

Note that while unfiltering any message is a possible security risk,
1) A quick grep through the FLTK source code suggests it does not handle
   these events, so the only code that is touched is the new code in
   this commit.
2) The drag and drop code is very simple, and I have been careful in
   avoiding the chance of buffer overflows.
3) This is much safer than having a pipe to the backend, which is more
   complex and can more easily contain security exploits.

Resolves #15