CLEANUP: Use typed lists everywhere
Replace most uses of node_t and casts to/from various node values to
typed lists. This makes the API much more clear and less error prone.

Note that qqueue still uses node_t directly, as it already provides an
interface with types.

While the names of lists might change in the near future because
they are currently on the longer side, it would be a simple find and
replace to rename them.

Resolves #7