FEATURE: Change strings to suggest installing from CD drives
The one remaining reference to ISOs is the drop area. It looks like
FLTK doesn't support a file and directory select window. I believe
the simplest solution (until the GUI redesign) is to either have two
tabs, or two drop areas on one tab. (Both will accept dropping ISOs
and drives, but the text and onclick handlers would be different).

For #20